Submit job execution results to broker from a single thread
Worker threads used to be responsible for submitting the result
of the execution of jobs to the broker. This meant that in the worse
case scenario one connection to redis could be open for each worker
thread. Now only a single connection is needed for the entire worker
pool.

This modification also opens the door to submitting results to the
broker in batches.