Only generate as many UUIDs as there are periodic tasks
Lua cannot generate random UUIDs so they are generated in Python and
passed to the `move_future_jobs` lua script. This script is responsible
for creating jobs when periodic tasks are due.

There used to be an arbitrary limit of 10 jobs that could be created
out of periodic tasks per call of the script. This commit generates as
many UUIDs as there are periodic tasks.

This removes the arbitrary limit and prevents generating useless UUIDs
if an application does not have many periodic tasks.