Switch back to sending signals from worker threads
Commit cc9a344738a50 created the notifier thread that receives results
of job executions and reports them to the broker. During the change,
post-run signals moved from being executed by workers to being
executed by this notifier thread.

This has the consequence to break raven and probably other things
depending on workers sending signals.

This commit switches back to sending signals from worker threads while
still keeping the notifier thread for notifying the broker. The idea
being that in the future calls to the broker can be batched to improve
throughput.