Allow to safely retry enqueuing jobs
Add an indempotency token to Redis scripts that enqueue jobs. This
allows a client to execute the script multiple times while enqueuing
jobs only once.

This is needed in case of a timeout after the command has been sent to
Redis, as it is impossible to know if the Redis server has processed the
command or not.