Allow to re-enqueue running jobs of a dead worker
Currently, when a worker gets killed and cannot exit gracefully,
it leaves a data-structure in Redis containing the jobs that it
was running before dying.

This commit adds a method to all brokers allowing to leverage
this data-structure to re-enqueue the jobs when a worker was found
dead.

Detecting that a worker is dead is not currently implemented,
relying on human intervention to mark it as dead and re-enqueue
its jobs.