Make the Redis broker clean up aborted jobs
The Redis broker uses a non-zero value for `max_retries` to determine
if it should remove a job from the "running" state.  So when an
`AbortException` is thrown, max out the `job.retries` counter,
rather than zeroing out the `job.max_retries` counter.  This will
have the same effect, but play more nicely with what the Redis
broker is expecting.

The MemoryBroker is unaffected by this.

Fixes Issue #20