Make Flask extension compatible with Flask>=2.2.0
Flask 2.2.0 deprecated the context stack in favour of using Python
ContextVar, and introduced a fake stack for some sort of backward
compatibility. However, it is not compatible enough for us as the
existing code would blindly push context replacing any existing
one, which will break application code.

The extension now does the recommended action of storing its data on
flask.g - in this case we store the context that needs to be popped
since _app_ctx_stack is deprecated.

There are no existing tests for this extension but I've tested it
locally to my satisfaction.

Fixes: https://github.com/NicolasLM/spinach/issues/24